Create syscallDefinition.diff
I did some research about log systems and syscall tables and did some digging around in 
reptilian-kernel. I found the syscall table in the file syscall_64.tbl and added my own get
and set functions at the end, starting with number 335. I made a process_log_level int variable 
and wrote the get and set functions in sys.c because the website I linked in the references 
portion of the man file said that is where they go. I went to office hours because I wasn’t
exactly sure how I was supposed to write the get and set system calls and it was there that 
I learned about the current_uid().val for checking if the user is a superuser. When making 
add_message I first checked if the level inputted was between 0 and 7 and returned -1 if not. 
I made a switch with cases for log levels 0-7 with their corresponding process level names and
the message passed in as an argument in the function.

	For process_log.c and .h I defined the functions, made the get and set functions return 
their corresponding syscalls, and defined PROC_LOG_CALL as 337. The retrieve set and get level
params returned an int pointer that I used malloc to make. Set included the syscall number, the
number of parameters and the new level, and get included the syscall number at 0 and the number
of parameters(0) at 1.